% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functioncode.R, R/multiplot.R
\name{multiplot}
\alias{multiplot}
\title{Title}
\usage{
multiplot(..., plotlist = NULL, file, cols = 1, layout = NULL)

multiplot(..., plotlist = NULL, file, cols = 1, layout = NULL)
}
\arguments{
\item{layout}{}
}
\value{

}
\description{
Title

Title
}
