% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyAppServer.R
\name{shinyAppServer}
\alias{shinyAppServer}
\title{Shiny app server function}
\usage{
shinyAppServer(input, output)
}
\arguments{
\item{input}{}

\item{output}{}
}
\description{
Shiny app server function
}
